layout/generic/test/test_bug670508.html

<!DOCTYPE>
<html>
<head>
<title>test for bug 670508</title>
<script type="text/javascript"
  src="chrome://mochikit/content/MochiKit/packed.js"></script>
<script type="text/javascript"
  src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
<script type="text/javascript"
  src="chrome://mochikit/content/tests/SimpleTest/EventUtils.js"></script>
<link rel="stylesheet" type="text/css"
  href="chrome://mochikit/content/tests/SimpleTest/test.css" />
</head>
<body>
<table>
  <tr>
    <td id="td" contenteditable="true">
      here is an editable table cell
    </td>
  </tr>
</table>
<pre id="test">
<script class="testbody" type="text/javascript">
SimpleTest.waitForExplicitFinish();

function doTest() {
  synthesizeMouse(document.getElementById('td'), 5, 5, { ctrlKey: true });
  ok(true, "not crashed");
  SimpleTest.finish();
}

SimpleTest.waitForFocus(doTest, window);

</script>
</pre>
</body>
</html>
